Update prefix bytes encoding to include mh length
